Refactored Record/Extraction/HashMapper
